Added Party Mode; Fixed Groundtruth Image bug
Added Party Mode - will sweep the main lights through the color wheel
at the pulse rate.

Fixed an implementation bug in the ground truth image gathering feature
- since we initially had all 648 bytes being returned at once, the
image gathering code was in the request event block. Since we have
changed it to a poke then 27 reads, that initial register with 0 in it
to read images off the sensors into memory was never triggered since it
was in the request event block.